+++ crates/ruff/src/rules/flake8_simplify/rules/ast_if.rs
@@ -9,8 +9,8 @@ use ruff_python_ast::helpers::{
     contains_call_path, contains_effect, create_expr, create_stmt, first_colon_range, has_comments,
     has_comments_in, unparse_expr, unparse_stmt,
 };
+use ruff_python_ast::newlines::UniversalNewlineIterator;
 use ruff_python_ast::types::Range;
-use ruff_python_ast::whitespace::UniversalNewlineIterator;
 
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
