+++ crates/ruff/src/rules/flake8_simplify/rules/ast_with.rs
@@ -5,8 +5,8 @@ use ruff_diagnostics::Diagnostic;
 use ruff_diagnostics::{AutofixKind, Availability, Violation};
 use ruff_macros::{derive_message_formats, violation};
 use ruff_python_ast::helpers::{first_colon_range, has_comments_in};
+use ruff_python_ast::newlines::UniversalNewlineIterator;
 use ruff_python_ast::types::Range;
-use ruff_python_ast::whitespace::UniversalNewlineIterator;
 
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
