+++ crates/ruff/src/rules/pycodestyle/rules/lambda_assignment.rs
@@ -3,9 +3,10 @@ use rustpython_parser::ast::{Arguments, Expr, ExprKind, Location, Stmt, StmtKind
 use ruff_diagnostics::{AutofixKind, Availability, Diagnostic, Fix, Violation};
 use ruff_macros::{derive_message_formats, violation};
 use ruff_python_ast::helpers::{match_leading_content, match_trailing_content, unparse_stmt};
+use ruff_python_ast::newlines::UniversalNewlineIterator;
 use ruff_python_ast::source_code::Stylist;
 use ruff_python_ast::types::{Range, ScopeKind};
-use ruff_python_ast::whitespace::{leading_space, UniversalNewlineIterator};
+use ruff_python_ast::whitespace::leading_space;
 
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
