+++ crates/ruff/src/rules/pydocstyle/rules/no_surrounding_whitespace.rs
@@ -1,8 +1,8 @@
 use ruff_diagnostics::{AlwaysAutofixableViolation, Diagnostic, Fix};
 use ruff_macros::{derive_message_formats, violation};
+use ruff_python_ast::newlines::NewlineWithTrailingNewline;
 use ruff_python_ast::str::leading_quote;
 use ruff_python_ast::types::Range;
-use ruff_python_ast::whitespace::NewlineWithTrailingNewline;
 
 use crate::checkers::ast::Checker;
 use crate::docstrings::definition::Docstring;
