+++ crates/ruff/src/rules/pydocstyle/rules/non_imperative_mood.rs
@@ -7,9 +7,9 @@ use ruff_diagnostics::{Diagnostic, Violation};
 use ruff_macros::{derive_message_formats, violation};
 use ruff_python_ast::cast;
 use ruff_python_ast::helpers::to_call_path;
+use ruff_python_ast::newlines::UniversalNewlineIterator;
 use ruff_python_ast::types::{CallPath, Range};
 use ruff_python_ast::visibility::{is_property, is_test};
-use ruff_python_ast::whitespace::UniversalNewlineIterator;
 
 use crate::checkers::ast::Checker;
 use crate::docstrings::definition::{DefinitionKind, Docstring};
